A_SetUserArray ACTORDEF function implementation
Note #1: If the variable does not exist we create a new one with the name given.
Note #2: No bounds check, since the array is stored as a sparse array.
Note #3: A variable and an array can share the same name
Note #4: The name is not nessesary to start with "user_"
(In ZDoom displays an error message)